pyproject.toml
[project]
name = "veadk-python"
version = "0.2.22"
description = "Volcengine agent development kit, integrations with Volcengine cloud services."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    {name = "Yaozheng Fang", email = "fangyozheng@gmail.com"},
    {name = "Guodong Li", email = "cu.eric.lee@gmail.com"},
    {name = "Zhi Han", email = "sliverydayday@gmail.com"},
    {name = "Meng Wang", email = "mengwangwm@gmail.com"}
]
dependencies = [
    "pydantic-settings>=2.10.1", # Config management
    "a2a-sdk>=0.3.0", # For Google Agent2Agent protocol
    "deprecated>=1.2.18",
    "google-adk>=1.10.0", # For basic agent architecture
    "litellm>=1.74.3", # For model inference
    "loguru>=0.7.3", # For better logging
    "opentelemetry-exporter-otlp>=1.35.0",
    "opentelemetry-instrumentation-logging>=0.56b0",
    "wrapt>=1.17.2", # For patching built-in functions
    "openai<1.100", # For fix https://github.com/BerriAI/litellm/issues/13710
    "volcengine-python-sdk>=4.0.33", # For Volcengine API
    "volcengine>=1.0.193", # For Volcengine sign
    "agent-pilot-sdk>=0.1.2", # Prompt optimization by Volcengine AgentPilot/PromptPilot toolkits
    "fastmcp>=2.11.3", # For running MCP
    "cookiecutter>=2.6.0", # For cloud deploy
    "omegaconf>=2.3.0", # For agent builder
    "llama-index>=0.14.0",
    "llama-index-embeddings-openai-like>=0.2.2",
    "llama-index-llms-openai-like>=0.5.1",
    "llama-index-vector-stores-opensearch>=0.6.1",
    "psycopg2-binary>=2.9.10", # For PostgreSQL database (short term memory)
    "pymysql>=1.1.1", # For MySQL database (short term memory)
    "opensearch-py==2.8.0",
    "filetype>=1.2.0",
    "trustedmcp>=0.0.2",
]

[project.scripts]
veadk = "veadk.cli.cli:veadk"

[project.optional-dependencies]
extensions = [
    "redis>=5.0",                 # For Redis database
    "tos>=2.8.4",                   # For TOS storage and Viking DB
    "llama-index-vector-stores-redis>=0.6.1",
]
database = [
    "redis>=5.0",                 # For Redis database
    "pymysql>=1.1.1",               # For MySQL database
    "volcengine>=1.0.193",          # For Viking DB
    "tos>=2.8.4",                   # For TOS storage and Viking DB
    "mem0ai==0.1.118",              # For mem0
]
speech = []
eval = [
    "prometheus-client>=0.22.1",    # For exporting data to Prometheus pushgateway
    "deepeval>=3.2.6",              # For DeepEval-based evaluation
    "google-adk[eval]",             # For Google ADK-based evaluation
]
cli = []
dev = [
    "pre-commit>=4.2.0",            # Format checking
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-xdist>=3.8.0",
]


[dependency-groups]
dev = [
    "pre-commit>=4.2.0",            # Format checking
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-xdist>=3.8.0",
]


[tool.setuptools.packages.find]
include = ["veadk*"]
exclude = ["assets*", "ide*", "tests*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"veadk" = ["**/*"]

[tool.ruff]
exclude = [
    "veadk/integrations/ve_faas/template/*",
    "veadk/integrations/ve_faas/web_template/*"
]
